+++ ui/src/state/chats.rs
@@ -1,0 +1,50 @@
+use std::collections::HashMap;
+
+use serde::{Deserialize, Serialize};
+use uuid::Uuid;
+use warp::raygun::Message;
+
+use super::identity::Identity;
+#[derive(Clone, Debug, PartialEq, Eq, Default, Deserialize, Serialize)]
+pub struct Chat {
+    // Warp generated UUID of the chat
+    // TODO: This should be wired up to warp conversation id's
+    #[serde(default)]
+    pub id: Uuid,
+    // Warp generated UUID of the chat
+    #[serde(default)]
+    pub active_media: bool, // TODO: in the future, this should probably be a vec of media streams or something
+    // Includes the list of participants within a given chat.
+    #[serde(default)]
+    pub participants: Vec<Identity>,
+    // Messages should only contain messages we want to render. Do not include the entire message history.
+    #[serde(default)]
+    pub messages: Vec<Message>,
+    // Unread count for this chat, should be cleared when we view the chat.
+    #[serde(default)]
+    pub unreads: u32,
+    // If a value exists, we will render the message we're replying to above the chatbar
+    #[serde(default)]
+    pub replying_to: Option<Message>,
+}
+
+// TODO: Properly wrap data which is expected to persist remotely in options, so we can know if we're still figuring out what exists "remotely", i.e. loading.
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+pub struct Chats {
+    // All active chats from warp.
+    #[serde(default)]
+    pub all: HashMap<Uuid, Chat>,
+    // Chat to display / interact with currently.
+    #[serde(default)]
+    pub active: Option<Uuid>,
+    // Chats to show in the sidebar
+    #[serde(default)]
+    pub in_sidebar: Vec<Uuid>,
+    // Favorite Chats
+    #[serde(default)]
+    pub favorites: Vec<Uuid>,
+}
+pub enum Direction {
+    Incoming,
+    Outgoing,
+}
