+++ ui/src/state/mod.rs
@@ -1,306 +1,56 @@
-use chrono::{DateTime, Utc};
+pub mod account;
+pub mod action;
+pub mod chats;
+pub mod friends;
+pub mod identity;
+pub mod route;
+pub mod settings;
+pub mod ui;
+
+// export specific structs which the UI expects. these structs used to be in src/state.rs, before state.rs was turned into the `state` folder
+pub use account::Account;
+pub use action::Action;
+pub use chats::{Chat, Chats};
+pub use friends::Friends;
+pub use identity::Identity;
+pub use route::Route;
+pub use settings::Settings;
+pub use ui::{Theme, ToastNotification, UI};
+
 use either::Either;
 use serde::{Deserialize, Serialize};
 use std::{
     collections::{BTreeMap, HashMap, HashSet},
-    fs,
-    hash::Hash,
+    fmt, fs,
 };
-
 use uuid::Uuid;
-use warp::{
-    constellation::item::Item,
-    crypto::DID,
-    multipass::identity::{Identity as WarpIdentity, IdentityStatus, Platform},
-    raygun::{Message, Reaction},
-};
-
-use kit::icons::Icon;
-
-#[derive(Eq, PartialEq)]
-pub struct MessageDivider {
-    pub timestamp: Option<DateTime<Utc>>,
-}
-
-impl Ord for MessageDivider {
-    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-        self.timestamp.cmp(&other.timestamp)
-    }
-}
-
-impl PartialOrd for MessageDivider {
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-pub enum Direction {
-    Incoming,
-    Outgoing,
-}
-
-#[derive(Clone, Debug, Default, Deserialize, Serialize, PartialEq)]
-pub struct ToastNotification {
-    pub title: String,
-    pub content: String,
-    #[serde(skip_serializing, skip_deserializing)]
-    pub icon: Option<Icon>,
-    initial_time: u32,
-    remaining_time: u32,
-}
-
-impl ToastNotification {
-    pub fn init(title: String, content: String, icon: Option<Icon>, timeout: u32) -> Self {
-        Self {
-            title,
-            content,
-            icon,
-            initial_time: timeout,
-            remaining_time: timeout,
-        }
-    }
-    pub fn remaining_time(&self) -> u32 {
-        self.remaining_time
-    }
-
-    pub fn reset_time(&mut self) {
-        self.remaining_time = self.initial_time
-    }
-
-    pub fn decrement_time(&mut self) {
-        if self.remaining_time > 0 {
-            self.remaining_time -= 1;
-        }
-    }
-}
-
-#[derive(PartialEq, Clone, Debug, Default, Deserialize, Serialize)]
-pub struct Theme {
-    pub filename: String,
-    pub name: String,
-    pub styles: String,
-}
-
-// Define a struct to represent a group of messages from the same sender.
-pub struct MessageGroup {
-    pub sender: DID,
-    pub remote: bool,
-    pub messages: Vec<GroupedMessage>,
-}
-
-// Define a struct to represent a message that has been placed into a group.
-pub struct GroupedMessage {
-    pub message: Message,
-    pub is_first: bool,
-    pub is_last: bool,
-}
-
-pub type Callback = Box<dyn Fn(&State, &Action)>;
-
-// Define a new struct to represent a hook that listens for a specific action type.
-pub struct ActionHook {
-    action_type: Either<Action, Vec<Action>>,
-    callback: Callback,
-}
-
-/// Alias for the type representing a route.
-pub type To = String;
-
-#[derive(Clone, Debug, Default, Deserialize, Serialize)]
-pub struct Account {
-    #[serde(default)]
-    pub identity: Identity,
-    // pub settings: Option<CustomSettings>,
-    // pub profile: Option<Profile>,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone, Eq)]
-pub struct Identity {
-    identity: WarpIdentity,
-    status: IdentityStatus,
-    platform: Platform,
-}
-
-impl Hash for Identity {
-    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
-        self.identity.hash(state)
-    }
-}
-
-impl PartialEq for Identity {
-    fn eq(&self, other: &Self) -> bool {
-        self.identity.eq(&other.identity)
-            && self.status.eq(&other.status)
-            && self.platform.eq(&other.platform)
-    }
-}
-
-impl Default for Identity {
-    fn default() -> Self {
-        Self::from(WarpIdentity::default())
-    }
-}
-
-impl From<WarpIdentity> for Identity {
-    fn from(identity: WarpIdentity) -> Self {
-        Identity {
-            identity,
-            status: IdentityStatus::Offline,
-            platform: Default::default(),
-        }
-    }
-}
-
-impl core::ops::Deref for Identity {
-    type Target = WarpIdentity;
-    fn deref(&self) -> &Self::Target {
-        &self.identity
-    }
-}
-
-impl core::ops::DerefMut for Identity {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.identity
-    }
-}
-
-impl Identity {
-    pub fn identity_status(&self) -> IdentityStatus {
-        self.status
-    }
-
-    pub fn platform(&self) -> Platform {
-        self.platform
-    }
-}
-
-impl Identity {
-    pub fn set_identity_status(&mut self, status: IdentityStatus) {
-        self.status = status;
-    }
-
-    pub fn set_platform(&mut self, platform: Platform) {
-        self.platform = platform;
-    }
-}
-
-#[derive(Clone, Debug, Default, Deserialize, Serialize)]
-pub struct Route {
-    // String representation of the current active route.
-    #[serde(default)]
-    pub active: To,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Default, Deserialize, Serialize)]
-pub struct Chat {
-    // Warp generated UUID of the chat
-    // TODO: This should be wired up to warp conversation id's
-    #[serde(default)]
-    pub id: Uuid,
-    // Warp generated UUID of the chat
-    #[serde(default)]
-    pub active_media: bool, // TODO: in the future, this should probably be a vec of media streams or something
-    // Includes the list of participants within a given chat.
-    #[serde(default)]
-    pub participants: Vec<Identity>,
-    // Messages should only contain messages we want to render. Do not include the entire message history.
-    #[serde(default)]
-    pub messages: Vec<Message>,
-    // Unread count for this chat, should be cleared when we view the chat.
-    #[serde(default)]
-    pub unreads: u32,
-    // If a value exists, we will render the message we're replying to above the chatbar
-    #[serde(default)]
-    pub replying_to: Option<Message>,
-}
-
-// TODO: Properly wrap data which is expected to persist remotely in options, so we can know if we're still figuring out what exists "remotely", i.e. loading.
-#[derive(Clone, Debug, Default, Deserialize, Serialize)]
-pub struct Chats {
-    // All active chats from warp.
-    #[serde(default)]
-    pub all: HashMap<Uuid, Chat>,
-    // Chat to display / interact with currently.
-    #[serde(default)]
-    pub active: Option<Uuid>,
-    // Chats to show in the sidebar
-    #[serde(default)]
-    pub in_sidebar: Vec<Uuid>,
-    // Favorite Chats
-    #[serde(default)]
-    pub favorites: Vec<Uuid>,
-}
-
-// TODO: Properly wrap data which is expected to persist remotely in options, so we can know if we're still figuring out what exists "remotely", i.e. loading.
-#[derive(Clone, Debug, Default, Deserialize, Serialize)]
-pub struct Friends {
-    // All active friends.
-    #[serde(default)]
-    pub all: HashMap<DID, Identity>,
-    // List of friends the user has blocked
-    #[serde(default)]
-    pub blocked: Vec<Identity>,
-    // Friend requests, incoming and outgoing.
-    #[serde(default)]
-    pub incoming_requests: Vec<Identity>,
-    #[serde(default)]
-    pub outgoing_requests: Vec<Identity>,
-}
-
-#[derive(Clone, Debug, Default, Deserialize, Serialize)]
-pub struct Files {
-    // All files
-    #[serde(default)]
-    pub all: Vec<Item>,
-    // Optional, active folder.
-    #[serde(default)]
-    pub active_folder: Option<Item>,
-}
-
-#[derive(Clone, Debug, Default, Deserialize, Serialize)]
-pub struct Settings {
-    // Selected Language
-    #[serde(default)]
-    pub language: String,
-}
-
-#[derive(Clone, Debug, Default, Deserialize, Serialize)]
-pub struct UI {
-    // Should the active video play in popout?
-    #[serde(default)]
-    pub popout_player: bool,
-    #[serde(default)]
-    pub muted: bool,
-    #[serde(default)]
-    pub silenced: bool,
-    #[serde(skip_serializing, skip_deserializing)]
-    pub toast_notifications: HashMap<Uuid, ToastNotification>,
-    #[serde(default)]
-    pub theme: Option<Theme>,
-    #[serde(default)]
-    pub enable_overlay: bool,
-}
-
-use std::fmt;
+use warp::{crypto::DID, raygun::Message};
 
 use crate::testing::mock::generate_mock;
 
+use self::{
+    action::ActionHook,
+    chats::Direction,
+    ui::{GroupedMessage, MessageGroup},
+};
+
+// todo: putting the State struct 300 lines into the file makes it hard to find :( state.rs should be turned into its own module and split into multiple files.
 #[derive(Default, Deserialize, Serialize)]
 pub struct State {
     #[serde(default)]
-    pub account: Account,
+    pub account: account::Account,
     #[serde(default)]
-    pub route: Route,
+    pub route: route::Route,
     #[serde(default)]
-    pub chats: Chats,
+    pub chats: chats::Chats,
     #[serde(default)]
-    pub friends: Friends,
+    pub friends: friends::Friends,
     #[serde(default)]
-    pub settings: Settings,
+    pub settings: settings::Settings,
     #[serde(default)]
-    pub ui: UI,
+    pub ui: ui::UI,
     #[serde(skip_serializing, skip_deserializing)]
-    pub(crate) hooks: Vec<ActionHook>,
+    pub(crate) hooks: Vec<action::ActionHook>,
 }
 
 impl fmt::Debug for State {
@@ -740,10 +490,10 @@ impl State {
     }
 
     pub fn clear(&mut self) {
-        self.chats = Chats::default();
-        self.friends = Friends::default();
-        self.account = Account::default();
-        self.settings = Settings::default();
+        self.chats = chats::Chats::default();
+        self.friends = friends::Friends::default();
+        self.account = account::Account::default();
+        self.settings = settings::Settings::default();
     }
 
     pub fn has_toasts(&self) -> bool {
@@ -917,82 +667,5 @@ impl State {
 
     pub fn mock() -> State {
         generate_mock()
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Deserialize, Serialize)]
-pub enum Action {
-    // UI
-    TogglePopout,
-    EndAll,
-    ToggleSilence,
-    ToggleMute,
-    SetOverlay(bool),
-    AddToastNotification(ToastNotification),
-    SetTheme(Theme),
-    ClearTheme,
-    // RemoveToastNotification,
-    ToggleMedia(Chat),
-    // Account
-    /// Sets the ID for the user.
-    SetId(Identity),
-
-    // Settings
-    /// Sets the selected language.
-    SetLanguage(String),
-
-    // Routes
-    /// Set the active route
-    Navigate(To),
-    // Requests
-    /// Send a new friend request
-    SendRequest(Identity),
-    /// To be fired when a friend request you sent is accepted
-    RequestAccepted(Identity),
-    /// Cancel an outgoing request
-    CancelRequest(Identity),
-
-    /// Handle a new incoming friend request
-    IncomingRequest(Identity),
-    /// Accept an incoming friend request
-    AcceptRequest(Identity),
-    /// Deny a incoming friend request
-    DenyRequest(Identity),
-
-    // Friends
-    RemoveFriend(Identity),
-    Block(Identity),
-    UnBlock(Identity),
-    /// Handles the display of "favorite" chats
-    Favorite(Chat),
-    UnFavorite(Chat),
-    /// Sets the active chat to a given chat
-    ChatWith(Chat),
-    /// Adds a chat to the sidebar
-    AddToSidebar(Chat),
-    /// Removes a chat from the sidebar, also removes the active chat if the chat being removed matches
-    RemoveFromSidebar(Chat),
-    /// Adds or removes a chat from the favorites page
-    ToggleFavorite(Chat),
-
-    // Messaging
-    /// Records a new message and plays associated notifications
-    NewMessage(Chat, Message),
-    /// React to a given message by ID
-    React(Chat, Message, Reaction),
-    /// Reply to a given message by ID
-    Reply(Chat, Message),
-    /// Prep the UI for a message reply.
-    StartReplying(Chat, Message),
-    /// Clears the reply for a given chat
-    CancelReply(Chat),
-    /// Sends a message to the given chat
-    Send(Chat, Message),
-    ClearUnreads(Chat),
-}
-
-impl Action {
-    fn compare_discriminant(&self, other: &Action) -> bool {
-        std::mem::discriminant(self) == std::mem::discriminant(other)
     }
 }
