+++ ui/src/state/settings.rs
@@ -1,0 +1,8 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Clone, Debug, Default, Deserialize, Serialize)]
+pub struct Settings {
+    // Selected Language
+    #[serde(default)]
+    pub language: String,
+}
